Updated CSS responsive website
