+++ .vscode/extensions.json
@@ -1,0 +1,5 @@
+{
+    "recommendations": [
+        "mohd-akram.vscode-html-format"
+    ]
+}
