+++ templates/Base.html
@@ -6,11 +6,13 @@
             {% endblock title%}
         </title>
         <link rel="shortcut icon" href="https://cdn1.iconfinder.com/data/icons/hawcons/32/699290-icon-60-file-ott-512.png" />
+        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css" rel="stylesheet">
+        <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"></script>
         <link rel="stylesheet" href="{{ url_for('static', filename='css/OTT_Service.css') }}">
         <script src="{{ url_for('static', filename='js/OTT_Service.js') }}"></script>
         <!--    <link rel="stylesheet" href="C:\Users\Admin\Desktop\DBMS\Project\Phase-2_19BCE148_19BCE149_19BCE151\static\css\OTT_Service.css"> -->
     </head>
-    <body>
+    <body style='background-color: rgb(44, 43, 43);'>
     {% block body %}
 
     {% endblock body%}
