+++ templates/OTT_Service_About_page.html
@@ -5,22 +5,28 @@
 {% block body %} 
 <div>
     <div style='background-color: rgb(44, 43, 43);'>
-    <center><img style="filter: brightness(150%);" src="https://cdn1.iconfinder.com/data/icons/hawcons/32/699290-icon-60-file-ott-512.png"  width="200px" height="75px"></center>
+    <center><img class="img-responsive" style="filter: brightness(150%);" src="https://cdn1.iconfinder.com/data/icons/hawcons/32/699290-icon-60-file-ott-512.png"  width="225px" height="125px"></center>
     </div>
-    <div id="div6">
-        <p style="font-family:cursive;font-size:25px;margin-left:50px;">
-            <br>
-            <span id="span2" style="color:red;font-size:30px;margin-left:0px;">What is OTT</span>
-            <br><br>An “over-the-top” media service is any online content provider that offers streaming media as a standalone product. The term is commonly applied to video-on-demand platforms, but also refers to audio streaming, messaging services, or internet-based voice calling solutions.
-            <br><br>High-value content at low cost: Streaming services are widely considered a cost-effective alternative to traditional cable packages. A Netflix HD subscription is currently $12.99 per month, while non-HD plans cost $8.99. That’s a reasonable fee even if you only intend to watch a fraction of its offered content.
-            <br><br>Original content: In recent years, OTT providers like Netflix and Amazon Prime have begun producing original content that is exclusively available through their service. Platforms like HBO Go and Disney+ also have exclusive streaming licenses for previously televised content.
-            <br><br>Compatibility with multiple devices: For years, watching cable television required a television set. Today, we can watch OTT content from a broad range of devices. Any account holder can enjoy the same OTT experience from a gaming console, smartphone, tablet, or smart TV.
-            <br><br>
-        </p>
-        <div id="div5" style="margin-top:0px;">
-        <p class='para'>
-            <img style="filter: brightness(150%);" src="https://cdn1.iconfinder.com/data/icons/hawcons/32/699290-icon-60-file-ott-512.png"  width="300px" height="150px">
-            <br>Terms and Privacy © 2020-2021, WWW.OTT_Service.com, Inc. or its affiliates<br><br><span style="font-size:25px;color:white">Created by 19BCE148-19BVE149-19BCE151</span></p>
+    <div class="position-relative shadow-1-strong rounded mb-5 opacity-70"
+        style="background-image: url('https://cdn.mos.cms.futurecdn.net/rDJegQJaCyGaYysj2g5XWY.jpg');
+               height: 80%;
+               background-repeat: no-repeat;
+               background-position: center;
+               background-size: cover;">
+                <div class="flex justify-content-center">
+                    <h1 class="span2"><br>What is OTT</h1>
+                    <p class="para2">
+                        <br>An “over-the-top” media service is any online content provider that offers streaming media as a standalone product. The term is commonly applied to video-on-demand platforms, but also refers to audio streaming, messaging services, or internet-based voice calling solutions.
+                        <br><br>High-value content at low cost: Streaming services are widely considered a cost-effective alternative to traditional cable packages. A Netflix HD subscription is currently $12.99 per month, while non-HD plans cost $8.99. That's a reasonable fee even if you only intend to watch a fraction of its offered content.
+                        <br><br>Original content: In recent years, OTT providers like Netflix and Amazon Prime have begun producing original content that is exclusively available through their service. Platforms like HBO Go and Disney+ also have exclusive streaming licenses for previously televised content.
+                        <br><br>Compatibility with multiple devices: For years, watching cable television required a television set. Today, we can watch OTT content from a broad range of devices. Any account holder can enjoy the same OTT experience from a gaming console, smartphone, tablet, or smart TV.
+                        <br><br>
+                    </p>
+                </div>
         </div>
-    </div>
+        <div id="div5" style="margin-top: -3rem;">
+            <p class='para'>
+            <img class="img-responsive" style="filter: brightness(150%);" src="https://cdn1.iconfinder.com/data/icons/hawcons/32/699290-icon-60-file-ott-512.png"  width="300px" height="150px">
+            <br>Terms and Privacy © 2020-2023, WWW.OTT_Service.com, Inc. or its affiliates<br><br><span style="font-size:125%;color:white">Feedback</span></p>
+        </div>
 {% endblock body %}
