+++ templates/OTT_Service_Forgotten_Password_Page.html
@@ -7,11 +7,23 @@
     <div>
     <center><img style="filter: brightness(150%);margin-top:10px;" src="https://cdn1.iconfinder.com/data/icons/hawcons/32/699290-icon-60-file-ott-512.png"  width="300px" height="150px"></center>
 </center></div>
-    <form style="margin-top:100px;margin-left: 150px;" action="/Set_password/" method="POST">
-    <div style="display:none"><input  type="text" id="Customer_id" name="Customer_id" value="{{id}}"></div>
-    <span class="span2">Email Ac.</span><span class="span2" style="margin-left:123px;">:</span><input style="margin-left:0px;" placeholder='Enter Your Email' class="btn8" type="email" id="email1" name="email1" value="{{values[0]}}"><span style="color: rgb(255, 81, 0);font-weight:bold;">{{error_e}}<br>
-    <span class="span2">New Password</span><span class="span2" style="margin-left:66.5px;">:</span><input style="margin-left:0px;" placeholder='Enter New Password' class="btn8" type="password" id="password1" name="password1" value="{{values[1]}}"><span style="color: rgb(255, 81, 0);font-weight:bold;">{{error_p}}<br>
-    <span class="span2">conform Passwoed</span><span class="span2" style="margin-left:21.5px;">:</span><input style="margin-left:0px;" placeholder='Enter New Password' class="btn8" type="password" id="password2" name="password2" value="{{values[2]}}"><span style="color: rgb(255, 81, 0);font-weight:bold;">{{error_r}}</span><br>
-    <button class="btn4" type="submit" style="font-size: 25px;margin-top:50px;margin-left:400px;">Set Password</button>
-</form>
+<div id="div8">
+    <div class="d-flex justify-content-center center">
+        <div id="div_log_in">
+            <br><br>
+            <form action="/Set_password/" method="POST">
+                <center><div style="display:none"><input  type="text" id="Customer_id" name="Customer_id" value="{{id}}"></div>
+                <span class="para1" style="margin-left:-2rem">Email Ac.</span><span style="margin-left:6.5rem" ></span><input  placeholder='Enter Your Email' class="btn8" type="email" id="email1" name="email1" value="{{values[0]}}"><br><br>
+                <span class="para1" style="margin-left:-2rem">New Password</span><span style="margin-left:2.5rem"></span><input  placeholder='Enter New Password' class="btn8" type="password" id="password1" name="password1" value="{{values[1]}}"><br><br>
+                <span class="para1" style="margin-left:-2rem">confirm Password</span><span class="para1"></span><input placeholder='Enter New Password' class="btn8" type="password" id="password2" name="password2" value="{{values[2]}}"><br><br>
+                <button class="btn4" type="submit" style="font-size: 25px;margin-top:50px;">Set Password</button><br>
+                <span style="color: rgb(255, 81, 0);font-weight:bold;">{{error_e}}<br></span><br>
+                <span style="color: rgb(255, 81, 0);font-weight:bold;">{{error_p}}<br></span><br>
+                <span style="color: rgb(255, 81, 0);font-weight:bold;">{{error_r}}<br></span><br></center>
+            </form>
+            <br>
+        </div>
+    </div>
+</div>
+
 {% endblock body %}
