+++ templates/OTT_Service_Login_page.html
@@ -7,25 +7,30 @@
         <div>
         <center><img style="filter: brightness(150%);margin-top:10px;" src="https://cdn1.iconfinder.com/data/icons/hawcons/32/699290-icon-60-file-ott-512.png"  width="300px" height="150px"></center>
     </center></div>
-        <div id="div_log_in">
-            <span class="span2" style="padding-left:0px;"><H1 style="text-align: center;">Sign in</H1></span>
-            <form action="/signed_in/" method="POST">
-            <span class="span2" style="margin-left: 75px;">Email Ac.</span><span class="span2" style="margin-left:0px;">:</span><input class="btn7" id="username" name="username" placeholder="Enter Email" type="text"><br><br>
-            <span class="span2" style="margin-left: 75px;">Password</span><span class="span2" style="margin-left:9px;">:</span><input id="singpass" name="singpass" placeholder="Enter Password" class="btn7"  type="password"><br><br><br>
-            <center><button style="height: 40px;width: 150px;padding-bottom:5px;" class="btn7" type="submit">Sign in</button></center></form>
-            {% if error_sing==1%}
-            <center><br><p>Invalid Ac. or Password</p></center>
-            {% endif %}
+    <div id="div8">
+        <div class="d-flex justify-content-center center">
+            <div id="div_log_in">
+                <span class="span2" style="padding-left:0px;"><H1 style="text-align: center;">Sign in</H1></span><br>
+                <form action="/signed_in/" method="POST">
+                <center><span class="para1">Email Ac.</span><span class="para1" style="margin-left:0px;">:</span><input class="btn7" id="username" name="username" placeholder="Enter Email" type="text"><br><br>
+                <span class="para1" >Password</span><span class="para1" style="margin-left:9px;">:</span><input id="singpass" name="singpass" placeholder="Enter Password" class="btn7"  type="password"><br><br>
+                <button class="btn4" type="submit">Sign in</button></center></form><br>
+                {% if error_sing==1%}
+                <center><br><p>Invalid Ac. or Password</p></center>
+                {% endif %}
+            </div>
         </div>
-        <br><br>
-        <div>
-            <form style="margin-left:530px;margin-top:0px;"action="/forgot_password/" method="POST">
+        <br>
+        <div class="d-flex justify-content-center center">
+            <form style="margin-top:2ren;"action="/forgot_password/" method="POST">
                 <div style="display:none"><input  type="text" id="Customer_id" name="Customer_id" value="{{id}}"></div>
-                <button style="margin-left:0px;height:40px;width: 150px;padding-bottom:5px;width: 250px;color:red;background-color:black" class="btn7" type="submit">Forgotten Password</button>
+                <button  class="btn4" type="submit">Forgotten Password</button>
             </form>
-            <form style="margin-left:530px;margin-top:10px;" action="/Sign_up/" method="POST">
+        </div>
+        <div class="d-flex justify-content-center center">
+            <form style="margin-top:1rem;" action="/Sign_up/" method="POST">
                 <div style="display:none"><input  type="text" id="Customer_id" name="Customer_id" value="{{id}}"></div>
-                <button type="submit" style="margin-left:0px;height:40px;width: 150px;padding-bottom:5px;width: 250px;color:red;background-color:black" class="btn7" type="submit">Sign up New_Account</button>
+                <button type="submit" class="btn4" type="submit">Sign up New_Account</button>
             </form>
         </div>
     </div>
